apps/requirements.txt
cassandra-driver==3.25.0
click==8.0.1
geomet==0.2.1.post1
importlib-metadata==4.0.1
pkg-resources==0.0.0
six==1.16.0
typing-extensions==3.10.0.0
zipp==3.4.1
